+++ judger2/task.py
@@ -5,7 +5,7 @@ from commons.task_typing import (Artifact, CompileResult, CompileTask, Input,
                                  InvalidTaskException, JudgeResult, JudgeTask, ResourceUsage,
                                  RunResult, StatusUpdateProgress, Testpoint,
                                  TestpointJudgeResult)
-from judger2.judger import ProgressReporter
+from judger2.interface import ProgressReporter
 from judger2.logging_ import task_logger
 from judger2.steps.check import check
 from judger2.steps.compile_ import compile
